func answer + test 3

+++ hw/vadim_maletski/test_hw4func.py
@@ -1,7 +1,7 @@
-from random import random  # noqa: DUO101
+from random import random  # noqa: DUO102
 
 
-def func_answer() -> int:
+def func_answer() -> float:
 
     ran = random()  # noqa: S311
     if ran < 0.5:
